requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
annotated-types==0.7.0
    # via pydantic
anyio==4.7.0
    # via
    #   sse-starlette
    #   starlette
    #   watchfiles
black==24.10.0
    # via -r requirements.in
cfgv==3.4.0
    # via pre-commit
click==8.1.8
    # via
    #   black
    #   uvicorn
distlib==0.3.9
    # via virtualenv
fastapi==0.115.6
    # via -r requirements.in
filelock==3.16.1
    # via virtualenv
greenlet==3.1.1
    # via -r requirements.in
h11==0.14.0
    # via uvicorn
identify==2.6.4
    # via pre-commit
idna==3.10
    # via anyio
iniconfig==2.0.0
    # via pytest
jinja2==3.1.5
    # via -r requirements.in
markupsafe==3.0.2
    # via jinja2
mypy==1.14.1
    # via -r requirements.in
mypy-extensions==1.0.0
    # via
    #   black
    #   mypy
nodeenv==1.9.1
    # via pre-commit
packaging==24.2
    # via
    #   black
    #   pytest
pathspec==0.12.1
    # via black
platformdirs==4.3.6
    # via
    #   black
    #   virtualenv
pluggy==1.5.0
    # via pytest
pre-commit==4.0.1
    # via -r requirements.in
pydantic==2.10.4
    # via
    #   -r requirements.in
    #   fastapi
pydantic-core==2.27.2
    # via pydantic
pytailwindcss==0.2.0
    # via -r requirements.in
pytest==8.3.4
    # via
    #   -r requirements.in
    #   pytest-asyncio
pytest-asyncio==0.25.1
    # via -r requirements.in
python-dotenv==1.0.1
    # via -r requirements.in
pyyaml==6.0.2
    # via pre-commit
ruff==0.8.5
    # via -r requirements.in
sniffio==1.3.1
    # via anyio
sse-starlette==2.2.1
    # via -r requirements.in
starlette==0.41.3
    # via
    #   -r requirements.in
    #   fastapi
    #   sse-starlette
tenacity==9.0.0
    # via -r requirements.in
typing-extensions==4.12.2
    # via
    #   anyio
    #   fastapi
    #   mypy
    #   pydantic
    #   pydantic-core
uvicorn==0.34.0
    # via -r requirements.in
virtualenv==20.28.0
    # via pre-commit
watchfiles==1.0.3
    # via -r requirements.in
